fix: update deployment status
